+++ cmake/XSProjects.cmake
@@ -28,6 +28,7 @@ add_library(xs_compiler
   xs/sources/driver/direct_xlil.c
   xs/sources/driver/options.c
   xs/sources/driver/source_native.c
+  xs/sources/driver/source_native_static.c
   xs/sources/driver/source_native_program.c
   xs/sources/driver/source_native_utils.c
   xs/sources/lexer.c