+++ tests/fixtures/source/MatchPatternTypeMismatch.xs
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
+// SPDX-License-Identifier: Apache-2.0
+
+fn main() -> Long {
+  value: Long = 2;
+  match (value) {
+    true -> {
+      return 7;
+    },
+    else -> {
+      return 0;
+    },
+  }
+  return 0;
+}